user_service_v2/app/db.py
from pymongo import MongoClient
import os
from dotenv import load_dotenv

load_dotenv()

MONGO_URL = os.getenv("MONGO_URL")

client = MongoClient(
    MONGO_URL,
    tls=True,
    tlsAllowInvalidCertificates=True
)

# V2 → writes to user_db.users_v2
db = client["user_db"]
users_collection = db["users_v2"]
