user_service_v2/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
pydantic[email]
email-validator
